JLT-165233 Install and configure host agent with puppet. - do not create link if already exists

+++ manifests/role/host_agent.pp
@@ -68,6 +68,8 @@ class dynatrace::role::host_agent (
   exec {"Creates link to execute service":
     command => "ln -s ${host_installer_prefix_dir}/dynatrace/init.d/dynaTraceHostagent /etc/init.d/${service}",
     path    => ['/usr/bin', '/usr/sbin', '/bin'],
+    unless  => ["test -L /etc/init.d/${service}"],
+    }
   } ->
   service { "Start and enable the ${role_name}'s service: '${service}'":
     ensure => $service_ensure,
